ls and cd now work with ext2
I implemented the methods:

ext2_findFileInDirectory, which returns the inode of a file/directory in the given directory (or 0 if none)

and methods to list a directory by inode.

It is worth mentioning that cd must use relative paths, and you cannot use slashes. So, you can only change to a directory that is inside of the current directory.