ext2 file reading, working towards vfs and elf
Now files can be read with ext2 (for now, you can print the contents using cat in the terminal.) Because of the kernel heap space limit, you can only use cat for up to 1MiB or so. But, the file reader has a theoretical limit of 16MB I think, because it uses the singly and doubly indirect pointers (though not the triply yet.)

Also, I am working towards making a vfs system, where you just have an ambiguous filesystem_t that has methods like read_file, probe, etc, so that way the implementation reading and writing files won't depend on the filesystem.

Also, I'm working on elf loading.