Raw executables (up to 4KiB) can be loaded and executed.
It uses the same system as before with the fat32 raw executables, except this time it runs it as a new process (and the shell waits for it.)

Or, you can use the bg command to create a new process and not wait for it, or the dummy command to create a dummy process. (Just a while(1))

It doesn't check the executable flag or anything yet.